refactor(context): move OScriptModuleTypeResolver into context
The resolver is a workspace-scoped URI->ModuleType registry that DocumentContext
reads to determine a document's module type. It depended on no types.* code
(only the infrastructure.WorkspaceScope leaf and the external mdclasses ModuleType),
yet living in types.oscript it was the sole context->types edge.

Relocating it to context removes that edge: context no longer imports types for
module-type resolution. types.oscript.OScriptLibraryIndex stays the producer that
registers module types into the context-owned registry (types->context, the
natural direction).
